add PR comments
